web/requirements.txt
flask
python-dotenv
flask-restful
requests
flask_breadcrumbs